clang-cross: workaround missing clang-native when using sstate cache
For some reason on Yocto 2.1 clang-native is not installed into sysroot
when clang-cross-${TARGET_ARCH} is installed, although it's marked as
its dependency via DEPENDS variable. On OE master it works as expected.

A long term goal would be to determine why it doesn't work as expected
and provide a real fix.